requirements.txt
pymongo
requests
apache-airflow[ssh]
bs4
google.cloud
slack_sdk
